app/requirements.txt
torch==2.0.1
transformers==4.33.3
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
